streamlit/requirements.txt
aiortc==1.3.2
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
seaborn==0.11.2
streamlit_option_menu==0.3.2
streamlit_webrtc==0.43.0
tensorflow==2.9.1